Refactor tests and apply the same fixes to Spark.
