+++ spark/v3.5/spark/src/main/java/org/apache/iceberg/spark/SparkUtil.java
@@ -287,7 +287,7 @@ public class SparkUtil {
    * @param value a value that is an instance of {@link Type.TypeID#javaClass()}
    * @return the value converted for Spark
    */
-  public static Object convertConstant(Type type, Object value) {
+  public static Object internalToSpark(Type type, Object value) {
     if (value == null) {
       return null;
     }
@@ -295,6 +295,7 @@ public class SparkUtil {
     switch (type.typeId()) {
       case DECIMAL:
         return Decimal.apply((BigDecimal) value);
+      case UUID:
       case STRING:
         if (value instanceof Utf8) {
           Utf8 utf8 = (Utf8) value;
@@ -325,7 +326,7 @@ public class SparkUtil {
           Types.NestedField field = fields.get(index);
           Type fieldType = field.type();
           values[index] =
-              convertConstant(fieldType, struct.get(index, fieldType.typeId().javaClass()));
+              internalToSpark(fieldType, struct.get(index, fieldType.typeId().javaClass()));
         }
 
         return new GenericInternalRow(values);
