+++ spark/v3.5/spark/src/main/java/org/apache/iceberg/spark/source/BaseReader.java
@@ -184,9 +184,9 @@ abstract class BaseReader<T, TaskT extends ScanTask> implements Closeable {
   protected Map<Integer, ?> constantsMap(ContentScanTask<?> task, Schema readSchema) {
     if (readSchema.findField(MetadataColumns.PARTITION_COLUMN_ID) != null) {
       StructType partitionType = Partitioning.partitionType(table);
-      return PartitionUtil.constantsMap(task, partitionType, SparkUtil::convertConstant);
+      return PartitionUtil.constantsMap(task, partitionType, SparkUtil::internalToSpark);
     } else {
-      return PartitionUtil.constantsMap(task, SparkUtil::convertConstant);
+      return PartitionUtil.constantsMap(task, SparkUtil::internalToSpark);
     }
   }
